Add confirmation dialog and popup controls

+++ src/lightnap-ng/src/app/admin/components/pages/role/role.component.ts
@@ -1,23 +1,35 @@
 import { AdminService } from "@admin/services/admin.service";
 import { CommonModule } from "@angular/common";
 import { Component, inject, Input, OnInit } from "@angular/core";
-import { ApiResponse, RoutePipe, SuccessApiResponse } from "@core";
+import { RouterLink } from "@angular/router";
+import { ApiResponse, ConfirmPopupComponent, RoutePipe, SuccessApiResponse } from "@core";
+import { ApiResponseComponent } from "@core/components/controls/api-response/api-response.component";
+import { ErrorListComponent } from "@core/components/controls/error-list/error-list.component";
+import { ConfirmationService } from "primeng/api";
+import { ButtonModule } from "primeng/button";
 import { CardModule } from "primeng/card";
+import { TableModule } from "primeng/table";
 import { combineLatest, map, Observable } from "rxjs";
 import { RoleViewModel } from "./role-view-model";
-import { RouterLink } from "@angular/router";
-import { TableModule } from "primeng/table";
-import { ButtonModule } from "primeng/button";
-import { ErrorListComponent } from "@core/components/controls/error-list/error-list.component";
-import { ApiResponseComponent } from "@core/components/controls/api-response/api-response.component";
 
 @Component({
   standalone: true,
   templateUrl: "./role.component.html",
-  imports: [CommonModule, CardModule, TableModule, ButtonModule, RouterLink, RoutePipe, ErrorListComponent, ApiResponseComponent],
+  imports: [
+    CommonModule,
+    CardModule,
+    TableModule,
+    ButtonModule,
+    RouterLink,
+    RoutePipe,
+    ErrorListComponent,
+    ApiResponseComponent,
+    ConfirmPopupComponent,
+  ],
 })
 export class RoleComponent implements OnInit {
   #adminService = inject(AdminService);
+  #confirmationService = inject(ConfirmationService);
 
   @Input() role!: string;
 
@@ -48,16 +60,25 @@ export class RoleComponent implements OnInit {
     );
   }
 
-  removeUserFromRole(userId: string) {
+  removeUserFromRole(event: any, userId: string) {
     this.errors = [];
-    this.#adminService.removeUserFromRole(userId, this.role).subscribe({
-      next: response => {
-        if (!response.result) {
-          this.errors = response.errorMessages;
-          return;
-        }
 
-        this.#refreshRole();
+    this.#confirmationService.confirm({
+        header: "Confirm Role Removal?",
+      message: `Are you sure that you want to remove this role membership?`,
+      target: event.target,
+      key: userId,
+      accept: () => {
+        this.#adminService.removeUserFromRole(userId, this.role).subscribe({
+          next: response => {
+            if (!response.result) {
+              this.errors = response.errorMessages;
+              return;
+            }
+
+            this.#refreshRole();
+          },
+        });
       },
     });
   }
